+++ partials/single/related-posts.php
@@ -184,7 +184,7 @@ if ( $oceanwp_related_query->have_posts() ) :
 					<?php endif; ?>
 
 					<h3 class="related-post-title">
-						<a href="<?php the_permalink(); ?>" title="<?php the_title_attribute(); ?>" rel="bookmark"><?php the_title(); ?></a>
+						<a href="<?php the_permalink(); ?>" rel="bookmark"><?php the_title(); ?></a>
 					</h3><!-- .related-post-title -->
 
 					<time class="published" datetime="<?php echo esc_html( get_the_date( 'c' ) ); ?>"><?php oceanwp_icon( 'date' ); ?><?php echo esc_html( get_the_date() ); ?></time>