+++ changelog.md
@@ -9,6 +9,7 @@
 * **Improved** - Accessibility: Header Replace search form style.
 * **Improved** - Accessibility: Header Overlay search form style.
 * **Improved** - Accessibility: Link Post Format, both entry and single.
+* **Improved** - Accessibility: Blog Entry titles.
 * **Fixed** - Previous item navigation arrow icon class.
 
 ### *2021.06.23.* - 2.1.1